# docs/swagger.yaml
definitions:
...
  DatabaseServicev1.Card:
    properties:
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      cvv:
        description: '* CVV код карты'
        type: integer
      date:
        description: '* Дата до которой активна карта'
        type: string
      fullName:
        description: '* ФИО с банковской карты'
        type: string
      id:
        description: '* ID банковской карты в базе данных'
        type: integer
      number:
        description: '* Номер карты'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
      userId:
        description: '* ID пользователя которому принадлежит данная карта'
        type: integer
    type: object
...
  DatabaseServicev1.CreateDonationsRequest:
    properties:
      amount:
        description: '* Сумма пожертвования'
        type: number
      title:
        description: '* Произвольное название пожертвования, примерно так: "На лекарства"'
        type: string
      userId:
        description: '* ID пользователя, чье это пожертвование'
        type: integer
      wardId:
        description: '* ID подопечный этого пожертвования'
        type: integer
    type: object
  DatabaseServicev1.CreateDonationsResponse:
    properties:
      amount:
        description: '* Сумма пожертвования'
        type: number
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      id:
        description: '* ID записи в базе данных'
        type: integer
      title:
        description: '* Название пожертвования (например "На лекарства")'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
      userId:
        description: '* ID пользователя, которому принадлежит пожертвование'
        type: integer
      wardId:
        description: '* ID подопечный этого пожертвования'
        type: integer
    type: object
...
  DatabaseServicev1.DeleteDonationByModelRequest:
    properties:
      amount:
        description: '* Сумма пожертвования'
        type: number
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      id:
        description: '* ID пожертвования в базе данных'
        type: integer
      title:
        description: '* Название пожертвования (например "На лекарства")'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
      userId:
        description: '* ID пользователя, которому принадлежит пожертвование'
        type: integer
      wardId:
        description: '* ID подопечный этого пожертвования'
        type: integer
    type: object
...
  DatabaseServicev1.Donations:
    properties:
      amount:
        description: '* Сумма пожертвования'
        type: number
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      id:
        description: '* ID пожертвования в базе данных'
        type: integer
      title:
        description: '* Название пожертвования (например "На лекарства")'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
      userId:
        description: '* ID пользователя, которому принадлежит пожертвование'
        type: integer
      wardId:
        description: '* ID подопечного для которого предназначено данное пожертвование'
        type: integer
    type: object
  DatabaseServicev1.DonationsResponse:
    properties:
      donations:
        description: '* Массив пожертвований'
        items:
          $ref: '#/definitions/DatabaseServicev1.Donations'
        type: array
    type: object
  DatabaseServicev1.FindDonationWardsResponse:
    properties:
      wards:
        allOf:
        - $ref: '#/definitions/DatabaseServicev1.Ward'
        description: '* Подопечный пожертвования'
    type: object
...
  DatabaseServicev1.HTTPCodes:
    properties:
      code:
        description: '* HTTP код ответа'
        type: integer
    type: object
...
  DatabaseServicev1.UpdateDonationsRequest:
    properties:
      amount:
        description: '* Сумма пожертвования'
        type: number
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      id:
        description: '* ID пожертвования в базе данных'
        type: integer
      title:
        description: '* Название пожертвования (например "На лекарства")'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
      userId:
        description: '* ID пользователя, которому принадлежит пожертвование'
        type: integer
      wardId:
        description: '* ID подопечный этого пожертвования'
        type: integer
    type: object
...
  DatabaseServicev1.Ward:
    properties:
      address:
        description: '* Адрес подопечного'
        type: string
      collected:
        description: '* Количество собранных средств'
        type: number
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      donations:
        description: '* пожертвования для подопечного'
        items:
          $ref: '#/definitions/DatabaseServicev1.Donations'
        type: array
      fullName:
        description: '* Полное имя подопечного'
        type: string
      id:
        description: '* ID подопечного в базе данных'
        type: integer
      necessary:
        description: '* Необходимая сумма денег на потребность'
        type: number
      title:
        description: '* Дополнительный текст к подопечному'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
      want:
        description: '* Потребность подопечного (то в чем он нуждается, например "Лекарства")'
        type: string
    type: object
...
  server.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: Сервер маршрутизации
  title: API Gateway
  version: "1.0"
paths:
...
  /api/v1/donations:
    get:
      consumes:
      - application/json
      description: Список всех пожертвований в базе данных
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.DonationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Список всех пожертвований
      tags:
      - Donations
    post:
      consumes:
      - application/json
      description: Создание пожертвования
      parameters:
      - description: Сущность пожертвования
        in: body
        name: donation
        schema:
          $ref: '#/definitions/DatabaseServicev1.CreateDonationsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.Card'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Создание пожертвования
      tags:
      - Donations
    put:
      consumes:
      - application/json
      description: Обновление пожертвования
      parameters:
      - description: Модель для обновления
        in: body
        name: donation
        schema:
          $ref: '#/definitions/DatabaseServicev1.UpdateDonationsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.CreateDonationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновление пожертвования
      tags:
      - Donations
  /api/v1/donations/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление пожертвования по ID
      parameters:
      - description: ID пожертвования
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.HTTPCodes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Удаление пожертвования
      tags:
      - Donations
    get:
      consumes:
      - application/json
      description: Поиск пожертвовани по ID
      parameters:
      - description: Donation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.CreateDonationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Поиск пожертвования
      tags:
      - Donations
  /api/v1/donations/{id}/user:
    get:
      consumes:
      - application/json
      description: Извлечение пользователя из пожертвования по ID пожертвования
      parameters:
      - description: Donation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.FindDonationWardsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Извлечение пользователя из пожертвования
      tags:
      - Donations
  /api/v1/donations/{id}/wards:
    get:
      consumes:
      - application/json
      description: Извлечение подопечных пожертвования по ID пожертвования
      parameters:
      - description: Donation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.FindDonationWardsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Извлечение подопечных пожертвования
      tags:
      - Donations
  /api/v1/donations/deleteModel:
    post:
      consumes:
      - application/json
      description: Удаляет пожертвование опираясь на всю сущность модели
      parameters:
      - description: Сущность пожертвования
        in: body
        name: donation
        schema:
          $ref: '#/definitions/DatabaseServicev1.DeleteDonationByModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.HTTPCodes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Удаление пожертвования по модели
      tags:
      - Donations
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"